requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
python-dotenv>=1.0.1
langchain>=0.2.14
langchain-community>=0.2.12
langchain-chroma>=0.1.2
langchain-huggingface>=0.0.3
langgraph>=0.2.14
chromadb>=0.5.5
sentence-transformers>=3.0.1
groq>=0.9.0
pypdf>=4.2.0
openpyxl>=3.1.5
